Move user feature field parsing logic to geofence class
