+++ app/src/main/AndroidManifest.xml
@@ -2,7 +2,6 @@
     package="com.ifttt.groceryexpress">
 
     <uses-permission android:name="android.permission.INTERNET"/>
-    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
 
     <application
         android:allowBackup="true"
