+++ connect-button/src/main/java/com/ifttt/connect/ui/ConnectButton.java
@@ -29,7 +29,6 @@ import com.ifttt.connect.ConnectionApiClient;
 import com.ifttt.connect.CredentialsProvider;
 import com.ifttt.connect.ErrorResponse;
 import com.ifttt.connect.R;
-import com.ifttt.connect.UserTokenAsyncTask;
 import com.ifttt.connect.api.PendingResult;
 
 import static android.animation.ValueAnimator.INFINITE;
