+++ connect-button/src/main/java/com/ifttt/connect/ui/UserTokenAsyncTask.java
@@ -1,6 +1,9 @@
-package com.ifttt.connect;
+package com.ifttt.connect.ui;
 
 import androidx.annotation.Nullable;
+import com.ifttt.connect.BuildConfig;
+import com.ifttt.connect.ConnectionApiClient;
+import com.ifttt.connect.CredentialsProvider;
 
 public final class UserTokenAsyncTask extends android.os.AsyncTask<Void, Void, String> {
