+++ connect-location/src/main/AndroidManifest.xml
@@ -1,6 +1,9 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.ifttt.location">
 
+    <uses-permission android:name="android.permission.INTERNET"/>
+    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
+
     <application>
         <receiver android:name="com.ifttt.location.AwarenessEnterReceiver" />
         <receiver android:name="com.ifttt.location.AwarenessExitReceiver" />
