+++ connect-location/src/main/java/com/ifttt/location/GeofenceProvider.java
@@ -1,10 +1,11 @@
 package com.ifttt.location;
 
+import com.ifttt.connect.Feature;
 import com.ifttt.connect.UserFeatureField;
 import java.util.List;
 
 interface GeofenceProvider {
 
-    void updateGeofences(final List<UserFeatureField> userFeatureFields);
+    void updateGeofences(final List<Feature> features);
 
 }
